Problem 7
PROBLEM STATEMENT---->
monkey starts riding a pole of "Z" meter height. It rides at the speed of "X" meter per minute. After a ride of one minute, it has to take rest for the next minute, during which it slips by "Y" meter down. It keeps going up and down in the same manner until it reaches at the top. Once it reach at the top of the pole it won't slide. Assume Z., X & Y as integers. Write a C++ program to solve this problem